refactor(breadcrumb): remove items API, switch to ol, co-locate breadcrumb-item
Remove the comma-separated `items`/`label` attribute API — only the
slot-based pattern with `<lui-breadcrumb-item>` is supported going
forward, as it is the only one that allows real hrefs. Replace `<ul>`
with `<ol>` for correct ordered list semantics. Move breadcrumb-item.ts
into the breadcrumb folder and update the index import path.
