+++ packages/lets-ui-components/src/components/breadcrumb/Breadcrumb.docs.mdx
@@ -1,9 +1,9 @@
-import { Meta, Canvas, Controls, ArgTypes } from "@storybook/addon-docs/blocks";
+import { Meta, Canvas, Controls, ArgTypes } from '@storybook/addon-docs/blocks';
 import * as BreadcrumbStories from './Breadcrumb.stories';
 
 <Meta of={BreadcrumbStories} />
 
 # Breadcrumb
 
-<Canvas of={BreadcrumbStories.Breadcrumb} />
-<Controls of={BreadcrumbStories.Breadcrumb} />
+<Canvas of={BreadcrumbStories.Default} />
+<Controls of={BreadcrumbStories.Default} />
