+++ packages/lets-ui-components/src/components/breadcrumb/Breadcrumb.stories.js
@@ -5,29 +5,25 @@ import '../../index.js';
 export default {
   title: 'Navigation/Breadcrumb',
   argTypes: {
-    label: { control: 'text' },
+    page: { control: 'text' },
+    link: { control: 'text' },
+    ariaLabel: { control: 'text' },
   },
 };
 
-const Template = ({ label }) =>
-  `<lui-breadcrumb items="Item 1,Item 2,Item 3" label="${label}"></lui-breadcrumb>`;
-
-export const Breadcrumb = Template.bind({});
-Breadcrumb.args = {
-  label: 'Page',
-};
-
-export const Primary = () => `
-  <lui-breadcrumb items="Item 1,Item 2,Item 3,Item 4" label="Item 5"></lui-breadcrumb>
-`;
-
-export const WithSlots = () => `
-  <lui-breadcrumb>
-    <lui-breadcrumb-item href="/item1">Item 1</lui-breadcrumb-item>
-    <lui-breadcrumb-item href="/item2">Item 2</lui-breadcrumb-item>
-    <lui-breadcrumb-item href="/item3">Item 3</lui-breadcrumb-item>
-    <lui-breadcrumb-item active>Página atual</lui-breadcrumb-item>
+const Template = ({ page, link, ariaLabel }) => `
+  <lui-breadcrumb aria-label="${ariaLabel}">
+    <lui-breadcrumb-item href="${link}">${page} 1</lui-breadcrumb-item>
+    <lui-breadcrumb-item href="${link}">${page} 2</lui-breadcrumb-item>
+    <lui-breadcrumb-item href="${link}">${page} 3</lui-breadcrumb-item>
+    <lui-breadcrumb-item active>${page} 4</lui-breadcrumb-item>
   </lui-breadcrumb>
 `;
-WithSlots.storyName = 'Com slots';
-WithSlots.parameters = { controls: { disable: true } };
+
+export const Default = Template.bind({});
+Default.parameters = { controls: { disable: true } };
+Default.args = {
+  page: 'Página',
+  link: '#',
+  ariaLabel: 'Breadcrumb',
+};
