+++ packages/lets-ui-components/src/components/breadcrumb/breadcrumb.scss
@@ -1,4 +1,5 @@
 @use 'packages/styles/src/components/breadcrumb' as *;
+@use 'packages/styles/src/components/link' as *;
 
 :host {
   display: block;
